Much refactoring

+++ src/com/zeejfps/sr/Camera.java
@@ -1,38 +1,38 @@
 package com.zeejfps.sr;
 
-import org.joml.Matrix4f;
-import org.joml.Quaternionf;
-import org.joml.Vector3f;
+import org.joml.Matrix4d;
+import org.joml.Quaterniond;
+import org.joml.Vector3d;
 
 public class Camera {
 
     public final Transform transform = new Transform();
 
-    private final Matrix4f viewMatrix = new Matrix4f();
-    private Matrix4f projMatrix;
+    private final Matrix4d viewMatrix = new Matrix4d();
+    private Matrix4d projMatrix;
 
     private float fov, aspect, zNear, zFar;
 
-    private Quaternionf realRotation = new Quaternionf();
+    private Quaterniond realRotation = new Quaterniond();
 
-    public final Vector3f forward = new Vector3f(0f, 0f, -1f);
-    public final Vector3f up = new Vector3f(0f, 1f, 0f);
+    public final Vector3d forward = new Vector3d(0f, 0f, -1f);
+    public final Vector3d up = new Vector3d(0f, 1f, 0f);
 
     public Camera(float fov, float aspect, float zNear, float zFar) {
         transform.position.z = -10f;
-        projMatrix = new Matrix4f()
+        projMatrix = new Matrix4d()
                 .perspective((float) Math.toRadians(fov), aspect, zNear, zFar);
     }
 
-    public Matrix4f getViewMatrix() {
+    public Matrix4d getViewMatrix() {
         realRotation.identity().rotateXYZ(transform.rotation.x, transform.rotation.y, transform.rotation.z);
         return viewMatrix.identity().lookAt(
                 transform.position,
-                forward.add(transform.position, new Vector3f()),
+                forward.add(transform.position, new Vector3d()),
                 up);
     }
 
-    public Matrix4f getProjMatrix() {
+    public Matrix4d getProjMatrix() {
         return projMatrix;
     }
