+++ src/com/zeejfps/sr/Mesh.java
@@ -1,18 +1,18 @@
 package com.zeejfps.sr;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 public class Mesh {
 
-    private final Vector3f[] vertices;
+    private final Vector3d[] vertices;
     private final int[] indecies;
 
-    public Mesh(Vector3f[] vertices, int[] indecies) {
+    public Mesh(Vector3d[] vertices, int[] indecies) {
         this.vertices = vertices;
         this.indecies = indecies;
     }
 
-    public Vector3f[] getVertices() {
+    public Vector3d[] getVertices() {
         return vertices;
     }
