+++ src/com/zeejfps/sr/Vertex.java
@@ -1,19 +1,19 @@
 package com.zeejfps.sr;
 
-import org.joml.Vector3f;
-import org.joml.Vector4f;
+import org.joml.Vector3d;
+import org.joml.Vector4d;
 
 public class Vertex {
 
-    public final Vector3f position;
+    public final Vector3d position;
     public final int color;
 
-    public Vertex(Vector3f position, int color) {
+    public Vertex(Vector3d position, int color) {
         this(position.x, position.y, position.z, color);
     }
 
-    public Vertex(float x, float y, float z, int color) {
-        this.position = new Vector3f(x, y, z);
+    public Vertex(double x, double y, double z, int color) {
+        this.position = new Vector3d(x, y, z);
         this.color = color;
     }
