+++ src/com/zeejfps/sr/utils/OBJImporter.java
@@ -1,7 +1,7 @@
 package com.zeejfps.sr.utils;
 
 import com.zeejfps.sr.Mesh;
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -12,7 +12,7 @@ public class OBJImporter {
 
     public static Mesh load(String path) throws IOException {
 
-        ArrayList<Vector3f> verts = new ArrayList<>();
+        ArrayList<Vector3d> verts = new ArrayList<>();
         ArrayList<Integer> indecies = new ArrayList<>();
 
         BufferedReader br = new BufferedReader(new FileReader(path));
@@ -26,7 +26,7 @@ public class OBJImporter {
                 float y = Float.parseFloat(tokens[2]);
                 float z = Float.parseFloat(tokens[3]);
 
-                verts.add(new Vector3f(x, y, z));
+                verts.add(new Vector3d(x, y, z));
 
             }
             else if (tokens[0].equals("f")){
@@ -53,7 +53,7 @@ public class OBJImporter {
 
         }
 
-        Mesh m = new Mesh(verts.toArray(new Vector3f[verts.size()]), indeci);
+        Mesh m = new Mesh(verts.toArray(new Vector3d[verts.size()]), indeci);
         return m;
 
     }
